pyproject.toml
[project]
name = "quclaw"
version = "0.1.0"
description = "Lightweight OpenClaw-style agent framework"
requires-python = ">=3.14"

dependencies = [
    "openai>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "httpx>=0.28.0",
    "watchdog>=3.0.0",
    "python-telegram-bot>=21.0",
    "discord.py>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]

[project.scripts]
quclaw = "cli.main:app"

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/channel",
    "src/cli",
    "src/core",
    "src/provider",
    "src/server",
    "src/tools",
    "src/utils",
]

[tool.hatch.build.targets.sdist]
include = [
    "src/channel",
    "src/cli",
    "src/core",
    "src/provider",
    "src/server",
    "src/tools",
    "src/utils",
    "pyproject.toml",
]
